// codeapp/src/generated/models/Pvci_inventorysyncrunsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pvci_inventorysyncrunsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pvci_inventorysyncrunsstatecode = keyof typeof Pvci_inventorysyncrunsstatecode;
export const Pvci_inventorysyncrunsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pvci_inventorysyncrunsstatuscode = keyof typeof Pvci_inventorysyncrunsstatuscode;

export interface Pvci_inventorysyncrunsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pvci_agentcount?: number;
  pvci_completedon?: string;
  pvci_createdcount?: number;
  pvci_environmentcount?: number;
  pvci_error?: string;
  pvci_inventorysyncrunid: string;
  pvci_name: string;
  pvci_rejectedcount?: number;
  pvci_runkey?: string;
  pvci_schemaversion?: string;
  pvci_source?: string;
  pvci_startedon?: string;
  pvci_status?: string;
  pvci_updatedcount?: number;
  statecode: Pvci_inventorysyncrunsstatecode;
  statuscode?: Pvci_inventorysyncrunsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pvci_inventorysyncruns extends Pvci_inventorysyncrunsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}